# apps/server/generated/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/study-sets": {
      "post": {
        "operationId": "StudySetController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStudySetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Study set created",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/StudySetResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a study set",
        "tags": [
          "Study Sets"
        ]
      }
    },
    "/study-sets/{id}": {
      "patch": {
        "operationId": "StudySetController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudySetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Study set updated",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/StudySetResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update own study set",
        "tags": [
          "Study Sets"
        ]
      },
      "get": {
        "operationId": "StudySetController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Study set retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/StudySetResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get study set details",
        "tags": [
          "Study Sets"
        ]
      },
      "delete": {
        "operationId": "StudySetController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Study set deleted",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete own study set",
        "tags": [
          "Study Sets"
        ]
      }
    },
    "/study-sets/{id}/visibility": {
      "patch": {
        "operationId": "StudySetController_updateVisibility",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVisibilityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Visibility updated",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/StudySetResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Make study set public/private",
        "tags": [
          "Study Sets"
        ]
      }
    },
    "/study-sets/me": {
      "get": {
        "operationId": "StudySetController_myStudySets",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Study sets retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/StudySetResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List study sets I created",
        "tags": [
          "Study Sets"
        ]
      }
    },
    "/study-sets/collection": {
      "get": {
        "description": "Owned + favourited sets, ordered by updatedAt desc. Pass ?q= to filter on title/description (case-insensitive substring).",
        "operationId": "StudySetController_myCollection",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Searches title or description",
            "schema": {
              "example": "biology",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Collection retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/StudySetResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List study sets in my collection",
        "tags": [
          "Study Sets"
        ]
      }
    },
    "/study-sets/public": {
      "get": {
        "operationId": "StudySetController_publicStudySets",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Searches title or description",
            "schema": {
              "example": "biology",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Public study sets retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/StudySetResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Search public study sets",
        "tags": [
          "Study Sets"
        ]
      }
    },
...
    "/folders/{id}/study-sets/{studySetId}": {
      "delete": {
        "operationId": "FolderStudySetController_removeStudySetFromFolder",
        "parameters": [
          {
            "name": "studySetId",
            "required": true,
            "in": "path",
            "description": "Study set id",
            "schema": {
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Folder id",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove a study set from folder",
        "tags": [
          "Folder Study Sets"
        ]
      }
    },
...
  },
  "info": {
    "title": "Mimir API",
    "description": "Auto-generated OpenAPI specification for the Mimir backend. This file is the source of truth for the frontend type pipeline.",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:9090",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt"
      }
    },
    "schemas": {
...
      "ResponseDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Operation successful"
          },
          "data": {
            "type": "object"
          },
          "error": {
            "type": "string"
          },
          "meta": {
            "type": "object"
          }
        },
        "required": [
          "ok",
          "message"
        ]
      },
      "StudySetResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "isPrivate": {
            "type": "boolean",
            "description": "true = private, false = public"
          },
          "ownerId": {
            "type": "string",
            "description": "Owner id"
          },
          "ownerName": {
            "type": "string"
          },
          "ownerImg": {
            "type": "string"
          },
          "isOwner": {
            "type": "boolean",
            "example": false
          },
          "isCollected": {
            "type": "boolean",
            "example": true
          },
          "flashcardsCount": {
            "type": "number",
            "example": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "description",
          "isPrivate",
          "ownerId",
          "isOwner",
          "flashcardsCount",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateStudySetDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Advanced Biology"
          },
          "description": {
            "type": "string",
            "example": "Notes and flashcards for bio exams"
          },
          "language": {
            "type": "string",
            "example": "German"
          },
          "isPrivate": {
            "type": "boolean",
            "example": false,
            "description": "true = private, false = public"
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateStudySetDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Advanced Biology"
          },
          "description": {
            "type": "string",
            "example": "Notes and flashcards for bio exams"
          },
          "language": {
            "type": "string",
            "example": "German"
          },
          "isPrivate": {
            "type": "boolean",
            "example": false,
            "description": "true = private, false = public"
          }
        }
      },
      "UpdateVisibilityDto": {
        "type": "object",
        "properties": {
          "isPrivate": {
            "type": "boolean",
            "example": false,
            "description": "Set to false to make the study set public"
          }
        },
        "required": [
          "isPrivate"
        ]
      },
...
    }
  }
}